feat(userserviceimpl): Agregado order

+++ src/main/java/com/nullers/restbookstore/rest/user/services/UserServiceImpl.java
@@ -1,5 +1,7 @@
 package com.nullers.restbookstore.rest.user.services;
 
+import com.nullers.restbookstore.rest.orders.models.Order;
+import com.nullers.restbookstore.rest.orders.repositories.OrderRepository;
 import com.nullers.restbookstore.rest.user.dto.UserInfoResponse;
 import com.nullers.restbookstore.rest.user.dto.UserRequest;
 import com.nullers.restbookstore.rest.user.dto.UserResponse;
@@ -34,18 +36,21 @@ import java.util.UUID;
 public class UserServiceImpl implements UserService {
     public static final String USER_NOT_FOUND_MSG = "Usuario no encontrado";
     private final UserRepository userRepository;
+    private final OrderRepository orderRepository;
     private final UserMapper userMapper;
     private final PasswordEncoder passwordEncode;
 
     /**
      * Constructor de la clase
      *
-     * @param userRepository repositorio de usuarios
-     * @param userMapper     mapper de usuarios
+     * @param userRepository  repositorio de usuarios
+     * @param orderRepository
+     * @param userMapper      mapper de usuarios
      * @param passwordEncode
      */
-    public UserServiceImpl(UserRepository userRepository, UserMapper userMapper, PasswordEncoder passwordEncode) {
+    public UserServiceImpl(UserRepository userRepository, OrderRepository orderRepository, UserMapper userMapper, PasswordEncoder passwordEncode) {
         this.userRepository = userRepository;
+        this.orderRepository = orderRepository;
         this.userMapper = userMapper;
         this.passwordEncode = passwordEncode;
     }
@@ -99,7 +104,8 @@ public class UserServiceImpl implements UserService {
     public UserInfoResponse findById(UUID id) {
         log.info("Buscando usuario por id: " + id);
         var user = userRepository.findById(id).orElseThrow(() -> new UserNotFound(USER_NOT_FOUND_MSG));
-        return userMapper.toUserInfoResponse(user);
+        var order = orderRepository.findOrderIdsByClientId(id).stream().map(p -> p.getId().toHexString()).toList();
+        return userMapper.toUserInfoResponse(user,order);
     }
 
     /**
@@ -168,6 +174,10 @@ public class UserServiceImpl implements UserService {
     public void deleteById(UUID id) {
         log.info("Borrando usuario por id: " + id);
         User user = userRepository.findById(id).orElseThrow(() -> new UserNotFound(USER_NOT_FOUND_MSG));
-        userRepository.delete(user);
+        if(orderRepository.existsByUserId(id)){
+            userRepository.updateIsDeletedToTrueById(id);
+        }else {
+            userRepository.delete(user);
+        }
     }
 }
